customization: set default e-steps to 138.2
Warning, this may not be the right value for your machine!
Elegoo's default firmware has e-steps at 130, which caused to be a slight
under-extrusion on my machine.

Check out how to calibrate the E-Steps here:
https://my3dlife.com/how-to-calibrate-your-e-steps-simple-guide-and-calculator/

Note that the LCD of the Neptune 3 does not support reading or setting this value.

Over OctoPrint or some other host, send M503 to read the current values (where it says
steps per unit), then use the above link to measure and calculate the new value.

Then go in and set the new value using M92 E<yournewlycalculatedvalue>.

If you want to make this setting permanent over printer power off/on cycles, write
the new settings to EEPROM using M500.