feat: enable advanced pause / filament change M600
Note that this requires a host (e.g. OctoPrint over USB). The  LCD
firmware does not know how to handle the necessary prompts.

For this to work go to your OctoPrint settings -> Printer -> Serial Connection ->
Firmware & Protocol and add M876 to the list of Emergency commands. The firmware
listens to this command for confirmation, and hence this need to be allowed
through by OctoPrint in the first place.

see https://marlinfw.org/docs/configuration/configuration.html#advanced-pause